test: refactor to use node test runner
