+++ .gitignore
@@ -9,3 +9,4 @@ docs/files
 /tests/browser/precompiled-templates.js
 /samples/express/js/govjucks.js
 /samples/express/js/templates.js
+junit.xml
