--- tests/loader.js
@@ -1,156 +1,0 @@
-(function() {
-  'use strict';
-
-  var expect,
-    Environment,
-    WebLoader,
-    FileSystemLoader,
-    NodeResolveLoader,
-    templatesPath;
-
-  if (typeof require !== 'undefined') {
-    expect = require('expect.js');
-    Environment = require('../src/environment').Environment;
-    WebLoader = require('../src/web-loaders').WebLoader;
-    FileSystemLoader = require('../src/node-loaders').FileSystemLoader;
-    NodeResolveLoader = require('../src/node-loaders').NodeResolveLoader;
-    templatesPath = 'tests/templates';
-  } else {
-    expect = window.expect;
-    Environment = govjucks.Environment;
-    WebLoader = govjucks.WebLoader;
-    FileSystemLoader = govjucks.FileSystemLoader;
-    NodeResolveLoader = govjucks.NodeResolveLoader;
-    templatesPath = '../templates';
-  }
-
-  describe('loader', function() {
-    it('should allow a simple loader to be created', function() {
-      // From Docs: http://mozilla.github.io/govjucks/api.html#writing-a-loader
-      // We should be able to create a loader that only exposes getSource
-      var env, parent;
-
-      function MyLoader() {
-        // configuration
-      }
-
-      MyLoader.prototype.getSource = function() {
-        return {
-          src: 'Hello World',
-          path: '/tmp/somewhere'
-        };
-      };
-
-      env = new Environment(new MyLoader(templatesPath));
-      parent = env.getTemplate('fake.njk');
-      expect(parent.render()).to.be('Hello World');
-    });
-
-    it('should catch loader error', function(done) {
-      // From Docs: http://mozilla.github.io/govjucks/api.html#writing-a-loader
-      // We should be able to create a loader that only exposes getSource
-      var env;
-
-      function MyLoader() {
-        // configuration
-        this.async = true;
-      }
-
-      MyLoader.prototype.getSource = function(s, cb) {
-        setTimeout(function() {
-          cb(new Error('test'));
-        }, 1);
-      };
-
-      env = new Environment(new MyLoader(templatesPath));
-      env.getTemplate('fake.njk', function(err, parent) {
-        expect(err).to.be.a(Error);
-        expect(parent).to.be(undefined);
-
-        done();
-      });
-    });
-
-    describe('WebLoader', function() {
-      it('should have default opts for WebLoader', function() {
-        var webLoader = new WebLoader(templatesPath);
-        expect(webLoader).to.be.a(WebLoader);
-        expect(webLoader.useCache).to.be(false);
-        expect(webLoader.async).to.be(false);
-      });
-
-      it('should emit a "load" event', function(done) {
-        var loader = new WebLoader(templatesPath);
-
-        if (typeof window === 'undefined') {
-          this.skip();
-        }
-
-        loader.on('load', function(name, source) {
-          expect(name).to.equal('simple-base.njk');
-          done();
-        });
-
-        loader.getSource('simple-base.njk');
-      });
-    });
-
-    if (typeof FileSystemLoader !== 'undefined') {
-      describe('FileSystemLoader', function() {
-        it('should have default opts', function() {
-          var loader = new FileSystemLoader(templatesPath);
-          expect(loader).to.be.a(FileSystemLoader);
-          expect(loader.noCache).to.be(false);
-        });
-
-        it('should emit a "load" event', function(done) {
-          var loader = new FileSystemLoader(templatesPath);
-          loader.on('load', function(name, source) {
-            expect(name).to.equal('simple-base.njk');
-            done();
-          });
-
-          loader.getSource('simple-base.njk');
-        });
-      });
-    }
-
-    if (typeof NodeResolveLoader !== 'undefined') {
-      describe('NodeResolveLoader', function() {
-        it('should have default opts', function() {
-          var loader = new NodeResolveLoader();
-          expect(loader).to.be.a(NodeResolveLoader);
-          expect(loader.noCache).to.be(false);
-        });
-
-        it('should emit a "load" event', function(done) {
-          var loader = new NodeResolveLoader();
-          loader.on('load', function(name, source) {
-            expect(name).to.equal('dummy-pkg/simple-template.html');
-            done();
-          });
-
-          loader.getSource('dummy-pkg/simple-template.html');
-        });
-
-        it('should render templates', function() {
-          var env = new Environment(new NodeResolveLoader());
-          var tmpl = env.getTemplate('dummy-pkg/simple-template.html');
-          expect(tmpl.render({foo: 'foo'})).to.be('foo');
-        });
-
-        it('should not allow directory traversal', function() {
-          var loader = new NodeResolveLoader();
-          var dummyPkgPath = require.resolve('dummy-pkg/simple-template.html');
-          expect(loader.getSource(dummyPkgPath)).to.be(null);
-        });
-
-        it('should return null if no match', function() {
-          var loader = new NodeResolveLoader();
-          var tmplName = 'dummy-pkg/does-not-exist.html';
-          expect(loader.getSource(tmplName)).to.be(null);
-        });
-      });
-    }
-  });
-}());
