--- tests/setup.js
@@ -1,2 +1,0 @@
-process.env.NODE_ENV = 'test';
-
